High-water mark buffers
This patch adds a set of functions for using “high-water mark”
buffers.  These buffers let you copy or append data without worrying
about the space used to store the data; the buffer will automatically
be reallocated when necessary to store all of the data.  It's called a
“high-water mark” buffer because it's possible for there to be less
data in the buffer than is allocated.  When using the buffer to store
a variable-sized field from a sequence of records, the buffer will end
up being large enough to hold the largest value of the field — the
high-water mark for that field.