Keep track of size of allocated memory
We weren't keeping track of the size of the buffer's allocated memory.
This would cause us to reallocate every time we copied data into the
buffer, even when there was already enough space.  We now also keep an
explicit count of the number of allocations and reallocations that are
performed.

Closes #1