Adding tests for allocation failures
There are now two test cases that test what happens when we can't
allocate enough memory.  They simulate this by trying to allocate
((size_t) -1) bytes, which should fail for all calls to malloc and
realloc.  We're currently doing the right thing for failed
allocations, which is good, but it's better to have the test!