Pointing at outside memory
HWM buffers can now point at outside memory.  This has a similar
effect as loading the memory into the buffer, but the data isn't
immediately copied into the buffer.  As long as you are accessing the
data in the buffer in a read-only fashion, no copying is required.
Once you perform an operation that requires modifying the buffer, the
outside data is copied in.  (Currently the “append” functions and the
functions that return a “writable” pointer force a copy.)