hwm_buffer_fprint function
We now have a function for printing out the contents of an HWM buffer.
This might not be especially useful in production systems, especially
if the buffers are large, but it's helpful for tests.  For instance,
our own test cases now print out the contents of the buffers when they
fail doing to mismatched data.