Reallocating now always updates data pointer
In hwm_buffer_ensure_size, we use realloc to reallocate our internal
buffer if it needs to be larger than it is.  This might involve moving
the buffer.  If it does, the pointer to the internal buffer is moved
correctly.

However, we've got separate pointers for the internal buffer and for
the actual data in the HWM buffer, since it might be pointing at
memory outside of the internal buffer.  That means that if the
reallocation moves the internal buffer, and the current data is in
that buffer, the data pointer isn't updated, too.

This was causing problems with the “append” methods: if the buffer is
pointing at outside memory before the append, they copy that data into
our own memory before performing the append.  If the data pointers
aren't updated correctly after resizing the buffer, and the buffer's
data is in our own memory, the “append” will be confused, and try to
copy over the data when it shouldn't.

The data pointer is now always updated correctly, fixing this issue.

Closes #2