Adding “scons sdist” target
This patch adds an “sdist” scons target for creating a source package.
Unforunately, the FindSourceFiles() scons function isn't very robust
right now, so we have to maintain the manifest of source files
manually.  The SConstruct file will also clean any sdist temporary
files when you run “scons -c”.  In the master branch, the version
number will be ethe next intended release, with a “-dev” suffix.  When
we merge over to release branches, we'll remove the suffix.